+++ src/main/java/com/clubtaekwondo/club/controller/admin/SubscriptionTypeController.java
@@ -1,8 +1,9 @@
 package com.clubtaekwondo.club.controller.admin;
 
-import com.clubtaekwondo.club.model.SubscriptionPeriod;
-import com.clubtaekwondo.club.model.SubscriptionType;
+import com.clubtaekwondo.club.model.*;
+import com.clubtaekwondo.club.service.SubscriptionService;
 import com.clubtaekwondo.club.service.SubscriptionTypeService;
+import com.clubtaekwondo.club.service.TariffService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -12,6 +13,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.context.request.WebRequest;
 
+import java.util.List;
 import java.util.Optional;
 
 @Controller
@@ -22,6 +24,10 @@ public class SubscriptionTypeController {
 
     @Autowired
     private SubscriptionTypeService subscriptionTypeService;
+    @Autowired
+    private SubscriptionService subscriptionService;
+    @Autowired
+    private TariffService tariffService;
 
     @GetMapping(value = "/typeList")
     public String typeList(Model model, WebRequest request) {
@@ -45,6 +51,7 @@ public class SubscriptionTypeController {
             return ("adminPart/type/addType");
         } else {
             subscriptionTypeService.save(type);
+            model.addAttribute("messageSuccess", " Le type d'abonnement a bien été ajouté. ");
         }
 
         model.addAttribute(TYPE, type);
@@ -72,6 +79,7 @@ public class SubscriptionTypeController {
             return ("adminPart/type/addType");
         } else {
             subscriptionTypeService.save(type);
+            model.addAttribute("messageSuccess", " Le type d'abonnement a bien été ajouté.");
         }
 
         model.addAttribute(TYPE, type);
@@ -84,7 +92,18 @@ public class SubscriptionTypeController {
     public String deleteType(@PathVariable("type") Long id, Model model) {
 
         SubscriptionType type = subscriptionTypeService.findById(id);
-        subscriptionTypeService.delete(type);
+        Optional<Subscription> firstSubscription = subscriptionService.getAllSubscription().stream().filter(s -> s.getSubscriptionType().getIdType().equals(type.getIdType()) && s.getSubscriptionStatus().equals(SubscriptionStatus.CONFIRMED)).findFirst();
+        if (firstSubscription.isPresent()) {
+            model.addAttribute("messageError", "Vous ne pouvez pas supprimer ce type d'abonnement ! Ce type d'abonnement est liée à des abonnements en cours.");
+        } else {
+            List<Tariff> tariffList = tariffService.getAllTariff();
+            for (Tariff tariff : tariffList) {
+                if (tariff.getTariffPK().getIdType().equals(type.getIdType())) {
+                    tariffService.delete(tariff);
+                }
+            }
+            subscriptionTypeService.delete(type);
+        }
 
         model.addAttribute("typeList", subscriptionTypeService.getAllSubscriptionType());
